son-monitor/manager/requirements.txt
Django==1.9.2
django-filter==0.12.0
django-rest-multiple-models==1.6.3
django-rest-swagger==0.3.5
djangorestframework==3.3.2
django-cors-headers==1.1.0
Markdown==2.6.5
mysqlclient==1.3.7
Pygments==2.1.1
PyYAML==3.11
psycopg2==2.6.2

